Update random-access-storage to 3.0.0
This bumps `random-access-storage` to 3.0.0, and fixes compilation
errors necessary on the process.

This commit implements `sync_all` using `Ok(())` as suggested on the PR
introducing the method as I've understood we are using `Vec` and they
are not buffered.

Related to https://github.com/datrs/random-access-storage/issues/6
Related to https://github.com/datrs/random-access-storage/pull/18
Related to https://github.com/datrs/random-access-storage/pull/17
Closes https://github.com/datrs/random-access-memory/pull/14